test(soft-rw): fix Python 3.10 contention
